move tag library to the place where it will be used
